+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Chg #154: Raise minimum required PHP version to 8.1 (@vjik)
 - Bug #155: Fix `Sort` configuration preparation (@vjik)
 - Bug #155: Fix same named order fields in `Sort` were not overriding previous ones (@vjik)
+- New #158: Add methods `PaginatorInterface::isSupportSorting()` and `PaginatorInterface::withSort()` (@vjik)
 
 ## 1.0.1 January 25, 2023